refactor(db): extract issuer subdomains and signing keys from organizations
Move the per-org issuer machinery out of db/organizations/mod.rs into
db/organizations/issuer.rs: subdomain claim/release (deterministic claim
slots, reuse cooldown, tombstones), find_org_by_subdomain,
any_subdomain_claimed, the org signing-key persistence
(deterministic_org_key_id, get/try_insert/list), SubdomainClaimError,
SUBDOMAIN_REUSE_COOLDOWN_SECS, and 15 tests.

The rotation-cancellation invariant is now owned by one file:
release_subdomain and cancel_org_rotation_in_tx live together, and the
latter loses its pub(crate) — nothing outside the module called it. The
auto-release hooks (subdomain_to_release, release_ineligible_subdomain)
become pub(super) for the sibling domain-lifecycle module; the
cross-module auto-release tests stay in mod.rs, which drops to 257
lines of org CRUD and module wiring. mod.rs re-exports the public
names, so db:: paths and all callers (services/oidc/org_keys.rs,
handlers, vouch-tests) are unchanged.

Closes #637
